Removed code duplication and magic numbers from test. Log entry verification generalized.
